+++ packages/sushi/src/chain-ids.ts
@@ -6,4 +6,5 @@ export const CHAIN_ID_ARRAY = [
   Chains.POLYGON_POS,
   Chains.BASE,
   Chains.ARBITRUM_ONE,
+  Chains.BLAST,
 ]