+++ packages/artblocks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rabbitholegg/questdk-plugin-artblocks
 
+## 1.0.0-alpha.4
+
+### Patch Changes
+
+- Updated dependencies [[`6e8f8e01`](https://github.com/rabbitholegg/questdk-plugins/commit/6e8f8e011f4251c6c179693d242dc6be34f76974)]:
+  - @rabbitholegg/questdk-plugin-utils@1.0.0-alpha.18
+
 ## 1.0.0-alpha.3
 
 ### Patch Changes